+++ ex3.h
@@ -31,9 +31,9 @@ struct Estado {
 };
 
 pair<vector<int>, vector<int>> dijkstraMetro(Planta*, int);
-vector<pair<pair<int, int>, double>> achaArestasMetro(Planta*, vector<int>);
-vector<pair<double, double>> calculaDistTempoCiclo(Planta*, vector<int>, int);
-vector<pair<pair<int, int>, pair<double, double>>> achaArestasOnibus(Planta*, vector<int>);
+vector<pair<pair<int, int>, int>> achaArestasMetro(Planta*, vector<int>);
+vector<pair<int, double>> calculaDistTempoCiclo(Planta*, vector<int>, int);
+vector<pair<pair<int, int>, pair<int, double>>> achaArestasOnibus(Planta*, vector<int>);
 PlantaBusca* constroiPlantaBusca(Planta*, vector<int>, Planta*, vector<int>);
 pair<double, double> calcula_custo_taxi(int, int, double, SegmentoBusca*);
 pair<double, double> calcula_custo(SegmentoBusca* atual, SegmentoBusca* adjacente, double distancia_taxi);
